Remove OrchardHash trait
