+++ src/bundle/commitments.rs
@@ -5,135 +5,20 @@ use blake2b_simd::{Hash as Blake2bHash, Params, State};
 use crate::{
     bundle::{Authorization, Authorized, Bundle},
     issuance::{IssueAuth, IssueBundle, Signed},
-    note_encryption::{OrchardDomainCommon, MEMO_SIZE},
-    orchard_flavor::{OrchardVanilla, OrchardZSA},
+    note_encryption::OrchardDomainCommon,
 };
 
-const ZCASH_ORCHARD_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrchardHash";
-const ZCASH_ORCHARD_ACTION_GROUPS_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcActGHash";
-const ZCASH_ORCHARD_ACTIONS_COMPACT_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcActCHash";
-const ZCASH_ORCHARD_ACTIONS_MEMOS_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcActMHash";
-const ZCASH_ORCHARD_ACTIONS_NONCOMPACT_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcActNHash";
+/// Personalization field for `orchard_digest`.
+pub const ZCASH_ORCHARD_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrchardHash";
 const ZCASH_ORCHARD_SIGS_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxAuthOrchaHash";
-const ZCASH_ORCHARD_ZSA_BURN_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcBurnHash";
 const ZCASH_ORCHARD_ZSA_ISSUE_PERSONALIZATION: &[u8; 16] = b"ZTxIdSAIssueHash";
 const ZCASH_ORCHARD_ZSA_ISSUE_ACTION_PERSONALIZATION: &[u8; 16] = b"ZTxIdIssuActHash";
 const ZCASH_ORCHARD_ZSA_ISSUE_NOTE_PERSONALIZATION: &[u8; 16] = b"ZTxIdIAcNoteHash";
 const ZCASH_ORCHARD_ZSA_ISSUE_SIG_PERSONALIZATION: &[u8; 16] = b"ZTxAuthZSAOrHash";
 
-fn hasher(personal: &[u8; 16]) -> State {
+/// Initialize a BLAKE2b hasher with the `personal` personalization.
+pub fn hasher(personal: &[u8; 16]) -> State {
     Params::new().hash_length(32).personal(personal).to_state()
-}
-
-/// Manages the evaluation of `orchard_digest`.
-pub trait OrchardHash {
-    /// OrchardDomain of the bundle (OrchardVanilla or OrchardZSA)
-    type OrchardDomain: OrchardDomainCommon;
-
-    /// Evaluate `orchard_digest` for the bundle as defined in
-    /// [ZIP-244: Transaction Identifier Non-Malleability][zip244]
-    /// for OrchardVanilla and as defined in
-    /// [ZIP-226: Transfer and Burn of Zcash Shielded Assets][zip226]
-    /// for OrchardZSA
-    ///
-    /// [zip244]: https://zips.z.cash/zip-0244
-    /// [zip226]: https://zips.z.cash/zip-0226
-    fn hash_bundle_txid_data<A: Authorization, V: Copy + Into<i64>>(
-        bundle: &Bundle<A, V, Self::OrchardDomain>,
-    ) -> Blake2bHash;
-
-    /// Incorporates the hash of actions (orchard_actions_compact_digest, orchard_actions_memos_digest,
-    /// orchard_actions_noncompact_digest) into the hasher.
-    fn update_hash_with_actions<A: Authorization, V: Copy + Into<i64>>(
-        main_hasher: &mut State,
-        bundle: &Bundle<A, V, Self::OrchardDomain>,
-    ) {
-        let mut ch = hasher(ZCASH_ORCHARD_ACTIONS_COMPACT_HASH_PERSONALIZATION);
-        let mut mh = hasher(ZCASH_ORCHARD_ACTIONS_MEMOS_HASH_PERSONALIZATION);
-        let mut nh = hasher(ZCASH_ORCHARD_ACTIONS_NONCOMPACT_HASH_PERSONALIZATION);
-
-        for action in bundle.actions().iter() {
-            ch.update(&action.nullifier().to_bytes());
-            ch.update(&action.cmx().to_bytes());
-            ch.update(&action.encrypted_note().epk_bytes);
-            ch.update(
-                &action.encrypted_note().enc_ciphertext.as_ref()
-                    [..Self::OrchardDomain::COMPACT_NOTE_SIZE],
-            );
-
-            mh.update(
-                &action.encrypted_note().enc_ciphertext.as_ref()
-                    [Self::OrchardDomain::COMPACT_NOTE_SIZE
-                        ..Self::OrchardDomain::COMPACT_NOTE_SIZE + MEMO_SIZE],
-            );
-
-            nh.update(&action.cv_net().to_bytes());
-            nh.update(&<[u8; 32]>::from(action.rk()));
-            nh.update(
-                &action.encrypted_note().enc_ciphertext.as_ref()
-                    [Self::OrchardDomain::COMPACT_NOTE_SIZE + MEMO_SIZE..],
-            );
-            nh.update(&action.encrypted_note().out_ciphertext);
-        }
-
-        main_hasher.update(ch.finalize().as_bytes());
-        main_hasher.update(mh.finalize().as_bytes());
-        main_hasher.update(nh.finalize().as_bytes());
-    }
-}
-
-impl OrchardHash for OrchardVanilla {
-    type OrchardDomain = OrchardVanilla;
-
-    /// Evaluate `orchard_digest` for the bundle as defined in
-    /// [ZIP-244: Transaction Identifier Non-Malleability][zip244]
-    ///
-    /// [zip244]: https://zips.z.cash/zip-0244
-    fn hash_bundle_txid_data<A: Authorization, V: Copy + Into<i64>>(
-        bundle: &Bundle<A, V, OrchardVanilla>,
-    ) -> Blake2bHash {
-        let mut h = hasher(ZCASH_ORCHARD_HASH_PERSONALIZATION);
-
-        Self::update_hash_with_actions(&mut h, bundle);
-
-        h.update(&[bundle.flags().to_byte()]);
-        h.update(&(*bundle.value_balance()).into().to_le_bytes());
-        h.update(&bundle.anchor().to_bytes());
-        h.finalize()
-    }
-}
-
-impl OrchardHash for OrchardZSA {
-    type OrchardDomain = OrchardZSA;
-
-    /// Evaluate `orchard_digest` for the bundle as defined in
-    /// [ZIP-226: Transfer and Burn of Zcash Shielded Assets][zip226]
-    ///
-    /// [zip226]: https://zips.z.cash/zip-0226
-    fn hash_bundle_txid_data<A: Authorization, V: Copy + Into<i64>>(
-        bundle: &Bundle<A, V, OrchardZSA>,
-    ) -> Blake2bHash {
-        let mut h = hasher(ZCASH_ORCHARD_HASH_PERSONALIZATION);
-        let mut agh = hasher(ZCASH_ORCHARD_ACTION_GROUPS_HASH_PERSONALIZATION);
-
-        Self::update_hash_with_actions(&mut agh, bundle);
-
-        agh.update(&[bundle.flags().to_byte()]);
-        agh.update(&bundle.anchor().to_bytes());
-        agh.update(&bundle.expiry_height().to_le_bytes());
-
-        h.update(agh.finalize().as_bytes());
-
-        let mut burn_hasher = hasher(ZCASH_ORCHARD_ZSA_BURN_HASH_PERSONALIZATION);
-        for burn_item in &bundle.burn {
-            burn_hasher.update(&burn_item.0.to_bytes());
-            burn_hasher.update(&burn_item.1.to_bytes());
-        }
-        h.update(burn_hasher.finalize().as_bytes());
-
-        h.update(&(*bundle.value_balance()).into().to_le_bytes());
-        h.finalize()
-    }
 }
 
 /// Evaluate `orchard_digest` for the bundle as defined in
@@ -147,7 +32,7 @@ impl OrchardHash for OrchardZSA {
 pub(crate) fn hash_bundle_txid_data<
     A: Authorization,
     V: Copy + Into<i64>,
-    D: OrchardDomainCommon + OrchardHash<OrchardDomain = D>,
+    D: OrchardDomainCommon,
 >(
     bundle: &Bundle<A, V, D>,
 ) -> Blake2bHash {
