+++ src/note_encryption/orchard_domain_vanilla.rs
@@ -1,10 +1,16 @@
 //! This module implements the note encryption logic specific for the `OrchardVanilla` flavor.
 
+use blake2b_simd::Hash as Blake2bHash;
 use zcash_note_encryption_zsa::note_bytes::NoteBytesData;
 
 use crate::{
+    bundle::{
+        commitments::{hasher, ZCASH_ORCHARD_HASH_PERSONALIZATION},
+        Authorization,
+    },
     note::{AssetBase, Note},
     orchard_flavor::OrchardVanilla,
+    Bundle,
 };
 
 use super::{
@@ -32,6 +38,23 @@ impl OrchardDomainCommon for OrchardVanilla {
 
     fn extract_asset(_plaintext: &Self::CompactNotePlaintextBytes) -> Option<AssetBase> {
         Some(AssetBase::native())
+    }
+
+    /// Evaluate `orchard_digest` for the bundle as defined in
+    /// [ZIP-244: Transaction Identifier Non-Malleability][zip244]
+    ///
+    /// [zip244]: https://zips.z.cash/zip-0244
+    fn hash_bundle_txid_data<A: Authorization, V: Copy + Into<i64>>(
+        bundle: &Bundle<A, V, OrchardVanilla>,
+    ) -> Blake2bHash {
+        let mut h = hasher(ZCASH_ORCHARD_HASH_PERSONALIZATION);
+
+        Self::update_hash_with_actions(&mut h, bundle);
+
+        h.update(&[bundle.flags().to_byte()]);
+        h.update(&(*bundle.value_balance()).into().to_le_bytes());
+        h.update(&bundle.anchor().to_bytes());
+        h.finalize()
     }
 }
