+++ src/note_encryption/orchard_domain_zsa.rs
@@ -1,10 +1,16 @@
 //! This module implements the note encryption logic specific for the `OrchardZSA` flavor.
 
+use blake2b_simd::Hash as Blake2bHash;
 use zcash_note_encryption_zsa::note_bytes::NoteBytesData;
 
 use crate::{
+    bundle::{
+        commitments::{hasher, ZCASH_ORCHARD_HASH_PERSONALIZATION},
+        Authorization,
+    },
     note::{AssetBase, Note},
     orchard_flavor::OrchardZSA,
+    Bundle,
 };
 
 use super::{
@@ -14,6 +20,9 @@ use super::{
     },
     orchard_domain::OrchardDomainCommon,
 };
+
+const ZCASH_ORCHARD_ACTION_GROUPS_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcActGHash";
+const ZCASH_ORCHARD_ZSA_BURN_HASH_PERSONALIZATION: &[u8; 16] = b"ZTxIdOrcBurnHash";
 
 impl OrchardDomainCommon for OrchardZSA {
     const COMPACT_NOTE_SIZE: usize = COMPACT_NOTE_SIZE_ZSA;
@@ -39,6 +48,35 @@ impl OrchardDomainCommon for OrchardZSA {
             .unwrap();
 
         AssetBase::from_bytes(bytes).into()
+    }
+
+    /// Evaluate `orchard_digest` for the bundle as defined in
+    /// [ZIP-226: Transfer and Burn of Zcash Shielded Assets][zip226]
+    ///
+    /// [zip226]: https://zips.z.cash/zip-0226
+    fn hash_bundle_txid_data<A: Authorization, V: Copy + Into<i64>>(
+        bundle: &Bundle<A, V, OrchardZSA>,
+    ) -> Blake2bHash {
+        let mut h = hasher(ZCASH_ORCHARD_HASH_PERSONALIZATION);
+        let mut agh = hasher(ZCASH_ORCHARD_ACTION_GROUPS_HASH_PERSONALIZATION);
+
+        Self::update_hash_with_actions(&mut agh, bundle);
+
+        agh.update(&[bundle.flags().to_byte()]);
+        agh.update(&bundle.anchor().to_bytes());
+        agh.update(&bundle.expiry_height().to_le_bytes());
+
+        h.update(agh.finalize().as_bytes());
+
+        let mut burn_hasher = hasher(ZCASH_ORCHARD_ZSA_BURN_HASH_PERSONALIZATION);
+        for burn_item in bundle.burn() {
+            burn_hasher.update(&burn_item.0.to_bytes());
+            burn_hasher.update(&burn_item.1.to_bytes());
+        }
+        h.update(burn_hasher.finalize().as_bytes());
+
+        h.update(&(*bundle.value_balance()).into().to_le_bytes());
+        h.finalize()
     }
 }
